Use uint64_t for RustBuffer len and cap
Task: LLTDEVOPS-510

v0.25.0 -> v0.28.0 broke how RustBuffer stores length and capacity,
now using uint64_t instead of int32_t.

This commit fixes this in go and c bridging code.

Part of basic test fixtures are now passing.

+++ bindgen/templates/BridgingHeaderTemplate.h
@@ -26,8 +26,8 @@
 // ⚠️ increment the version suffix in all instances of UNIFFI_SHARED_HEADER_V6 in this file.           ⚠️
 
 typedef struct RustBuffer {
-	int32_t capacity;
-	int32_t len;
+	uint64_t capacity;
+	uint64_t len;
 	uint8_t *data;
 } RustBuffer;
